reskin(T34+T35): canonicalize Strain and Sleep
Strain: route every presentation boundary through StrainScale on 0-21, keep constant blue, and align Today, detail, workouts, trends, widgets, Live Activity, watch, and complication copy.

Sleep C3 map: Today/Trends/widgets/notifications/watch show the score; Sleep hero shows score plus Duration, Efficiency, and Resting HR; need/debt remain in Sleep detail and Smart Alarm. Today Sleep now opens SleepView.

QA: stored 67 renders 14.1 on Today, detail, and Workouts. NOOPiOS and NOOPiOSWidgets build green. Watch smoke-build blocked because watchOS 26.5 runtime is not installed.

+++ StrandiOSShared/LiveActivityAttributes.swift
@@ -10,11 +10,11 @@ public struct NOOPActivityAttributes: ActivityAttributes {
         public var bpm: Int?
         public var recovery: Int?
         public var bonded: Bool
-        // Effort / strain on NOOP's 0–100 axis (#446) — one more stat in the Dynamic Island expanded
+        // Canonical 0–21 Strain display value (#446) — one more stat in the Dynamic Island expanded
         // region. OPTIONAL with a nil default so an activity started by an older build still decodes.
-        public var effort: Int?
+        public var effort: Double?
 
-        public init(bpm: Int?, recovery: Int?, bonded: Bool, effort: Int? = nil) {
+        public init(bpm: Int?, recovery: Int?, bonded: Bool, effort: Double? = nil) {
             self.bpm = bpm
             self.recovery = recovery
             self.bonded = bonded
