+++ StrandiOSShared/WidgetSnapshot.swift
@@ -11,13 +11,13 @@ public struct WidgetSnapshot: Codable, Equatable {
     public var updated: Date
     // Richer glance fields (#446). All OPTIONAL with nil defaults so a snapshot written by an OLDER app
     // build (which never encoded these keys) still decodes — Codable fills a missing optional with nil.
-    public var effort: Int?      // Effort / strain on NOOP's 0–100 axis
+    public var effort: Double?   // Strain on the canonical 0–21 display axis
     public var rest: Int?        // Rest (sleep_performance) score, 0–100
     public var hrv: Int?         // HRV (ms), whole-number for the glance
     public var restingHr: Int?   // Resting heart rate (bpm)
 
     public init(recovery: Int?, bpm: Int?, batteryPct: Int?, bonded: Bool, updated: Date,
-                effort: Int? = nil, rest: Int? = nil, hrv: Int? = nil, restingHr: Int? = nil) {
+                effort: Double? = nil, rest: Int? = nil, hrv: Int? = nil, restingHr: Int? = nil) {
         self.recovery = recovery
         self.bpm = bpm
         self.batteryPct = batteryPct
